[Rename] mySummary -> mypage
- 마이페이지에서 나오는 타입 전체 저장하는 파일로 변경

Resolves 마이페이지 개발 #53
